refactor time methods from anchore_engine.utils to anchore_engine.util.time
