+++ anchore_engine/plugins/authorization/client.py
@@ -6,7 +6,7 @@ from urllib import parse as urlparse
 import requests
 
 from anchore_engine.subsys import logger
-from anchore_engine.utils import datetime_to_rfc3339
+from anchore_engine.util.time import datetime_to_rfc3339
 
 
 class BasicApiClient(object):
