+++ anchore_engine/services/policy_engine/engine/policy/bundles.py
@@ -32,7 +32,7 @@ from anchore_engine.services.policy_engine.engine.policy.gates import *
 from anchore_engine.subsys import logger
 from anchore_engine.util.docker import parse_dockerimage_string
 from anchore_engine.util.matcher import is_match, regexify
-from anchore_engine.utils import rfc3339str_to_datetime
+from anchore_engine.util.time import rfc3339str_to_datetime
 
 
 class VersionedEntityMixin(object):
