+++ anchore_engine/utils.py
@@ -1,7 +1,6 @@
 """
 Generic utilities
 """
-import datetime
 import decimal
 import os
 import platform
@@ -380,68 +379,6 @@ def ensure_bytes(obj):
 
 def ensure_str(obj):
     return str(obj, "utf-8") if type(obj) != str else obj
-
-
-rfc3339_date_fmt = "%Y-%m-%dT%H:%M:%SZ"
-rfc3339_date_input_fmts = [
-    "%Y-%m-%dT%H:%M:%SZ",
-    "%Y-%m-%dT%H:%M:%S.%fZ",
-    "%Y-%m-%dT%H:%M:%S:%fZ",
-]
-
-
-def rfc3339str_to_epoch(rfc3339_str):
-    return int(rfc3339str_to_datetime(rfc3339_str).timestamp())
-
-
-def rfc3339str_to_datetime(rfc3339_str):
-    """
-    Convert the rfc3339 formatted string (UTC only) to a datatime object with tzinfo explicitly set to utc. Raises an exception if the parsing fails.
-
-    :param rfc3339_str:
-    :return:
-    """
-
-    ret = None
-    for fmt in rfc3339_date_input_fmts:
-        try:
-            ret = datetime.datetime.strptime(rfc3339_str, fmt)
-            # Force this since the formats we support are all utc formats, to support non-utc
-            if ret.tzinfo is None:
-                ret = ret.replace(tzinfo=datetime.timezone.utc)
-            continue
-        except:
-            pass
-
-    if ret is None:
-        raise Exception(
-            "could not convert input value ({}) into datetime using formats in {}".format(
-                rfc3339_str, rfc3339_date_input_fmts
-            )
-        )
-
-    return ret
-
-
-def datetime_to_rfc3339(dt_obj):
-    """
-    Simple utility function. Expects a UTC input, does no tz conversion
-
-    :param dt_obj:
-    :return:
-    """
-
-    return dt_obj.strftime(rfc3339_date_fmt)
-
-
-def epoch_to_rfc3339(epoch_int):
-    """
-    Convert an epoch int value to a RFC3339 datetime string
-
-    :param epoch_int:
-    :return:
-    """
-    return datetime_to_rfc3339(datetime.datetime.utcfromtimestamp(epoch_int))
 
 
 def convert_bytes_size(size_str):
