+++ tests/unit/anchore_engine/services/policy_engine/api/test_models.py
@@ -14,7 +14,7 @@ from anchore_engine.common.models.policy_engine import (
     LegacyTableReport,
     LegacyVulnerabilityReport,
 )
-from anchore_engine.utils import datetime_to_rfc3339
+from anchore_engine.util.time import datetime_to_rfc3339
 
 
 def test_feeds():
